Reconstructed ManageTradingInquiryActivity.
