--- app/src/main/res/layout/fragment_manage_trading_inquiry_today.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/background"
-    android:orientation="vertical">
-
-    <com.duang.mypinnedheaderlistview.PinnedHeaderListView
-        android:id="@+id/manage_trading_inquiry_today_list_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:groupIndicator="@null" />
-
-    <com.rey.material.widget.ProgressView
-        android:id="@+id/manage_trading_inquiry_today_progress_view"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="80dp"
-        android:visibility="gone"
-        app:pv_autostart="true"
-        app:pv_circular="true"
-        app:pv_progressMode="indeterminate"
-        app:pv_progressStyle="@style/Material.Drawable.CircularProgress" />
-
-    <ImageView
-        android:id="@+id/manage_trading_inquiry_today_nothing_founded_image_view"
-        android:layout_width="200dp"
-        android:layout_height="250dp"
-        android:layout_gravity="center"
-        android:visibility="visible" />
-
-</FrameLayout>
